info_sight/ocr_api/requirements.txt
google-genai
python-dotenv
fastapi
uvicorn
python-multipart
pillow
streamlit
pdf2image
pymupdf
